refactor: mv math.quantum.expand_vector -> math.matrix_manipulation.expand_vector
