+++ pennylane/math/matrix_manipulation.py
@@ -14,6 +14,7 @@
 """This module contains methods to expand the matrix representation of an operator
 to a higher hilbert space with re-ordered wires."""
 import itertools
+import numbers
 from collections.abc import Callable, Generator, Iterable
 from functools import reduce
 
@@ -348,3 +349,56 @@ def get_batch_size(tensor, expected_shape, expected_size):
             raise err
 
     return None
+
+
+def expand_vector(vector, original_wires, expanded_wires):
+    r"""Expand a vector to more wires.
+
+    Args:
+        vector (array): :math:`2^n` vector where n = len(original_wires).
+        original_wires (Sequence[int]): original wires of vector
+        expanded_wires (Union[Sequence[int], int]): expanded wires of vector, can be shuffled
+            If a single int m is given, corresponds to list(range(m))
+
+    Returns:
+        array: :math:`2^m` vector where m = len(expanded_wires).
+    """
+    if len(original_wires) == 0:
+        val = qml.math.squeeze(vector)
+        return val * qml.math.ones(2 ** len(expanded_wires))
+    if isinstance(expanded_wires, numbers.Integral):
+        expanded_wires = list(range(expanded_wires))
+
+    N = len(original_wires)
+    M = len(expanded_wires)
+    D = M - N
+
+    len_vector = qml.math.shape(vector)[0]
+    qudit_order = int(2 ** (np.log2(len_vector) / N))
+
+    if not set(expanded_wires).issuperset(original_wires):
+        raise ValueError("Invalid target subsystems provided in 'original_wires' argument.")
+
+    if qml.math.shape(vector) != (qudit_order**N,):
+        raise ValueError(f"Vector parameter must be of length {qudit_order}**len(original_wires)")
+
+    dims = [qudit_order] * N
+    tensor = qml.math.reshape(vector, dims)
+
+    if D > 0:
+        extra_dims = [qudit_order] * D
+        ones = qml.math.ones(qudit_order**D).reshape(extra_dims)
+        expanded_tensor = qml.math.tensordot(tensor, ones, axes=0)
+    else:
+        expanded_tensor = tensor
+
+    wire_indices = [expanded_wires.index(wire) for wire in original_wires]
+    wire_indices = np.array(wire_indices)
+
+    # Order tensor factors according to wires
+    original_indices = np.array(range(N))
+    expanded_tensor = qml.math.moveaxis(
+        expanded_tensor, tuple(original_indices), tuple(wire_indices)
+    )
+
+    return qml.math.reshape(expanded_tensor, (qudit_order**M,))
